First high-level refactor
